requirements.txt
anyio==3.6.1
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
db-dtypes==1.0.1
fastapi==0.78.0
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-cloud-pubsub==2.13.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpc-google-iam-v1==0.12.4
grpcio==1.46.3
grpcio-status==1.46.3
h11==0.13.0
httptools==0.4.0
idna==3.3
numpy==1.22.4
packaging==21.3
pandas==1.4.2
proto-plus==1.20.5
protobuf==3.20.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
